pyproject.toml
[tool.poetry]
name = "personal-stylist"
version = "0.1.0"

description = "Fashion assistant based on LLM with graph database"
authors = ["Daria Prudnikova"]
license = "MIT License"
readme = "README.md"
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "1.26.4"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
neo4j = "^5.28.1"
kafka-python = "^2.0.5"
redis = "^5.2.1"
torch = "^2.6.0"
transformers = "^4.49.0"
faiss-cpu = "^1.10.0"
chromadb = "^0.6.3"
aiogram = "^3.18.0"
peft = "^0.14.0"
loguru = "^0.7.3"
dvc = "^3.59.1"
prometheus-client = "^0.21.1"
python-dotenv = "^1.1.0"
openai = "^1.76.0"
aiohttp = "^3.11.18"
playwright = "^1.51.0"
lxml = "^5.4.0"
jsonlines = "^4.0.0"
datasets = "^2.16.0"
wandb = "0.15.12"
evaluate = "^0.4.0"
scikit-learn = "^1.4.0"
tqdm = "^4.66.0"
bert-score = "^0.3.13"
bitsandbytes = "^0.45.5"
deepspeed = "^0.16.7"
requests = "2.31.0"
sentencepiece = "^0.2.0"
nlpaug = "^1.1.11"
nltk = "^3.9.1"
pydantic-settings = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
